Overhaul WASM <-> JS data transfer
Introduce a better pattern for exchanging structured data (mainly
vectors, strings and structs) between WASM and JS. I'm sure this could be
simplified using something like wasm-bindgen, but I think its useful to
have a workable approach that requires minimal tooling (just a stock
Rust compiler with the wasm32-unknown-unknown target).

 - For transfering arrays between WASM and JS, allocate a `Box<Vec<T>>`
   on the Rust side and return a pointer to JS. This can be passed
   around in JS as an opaque pointer and passed back to WASM functions
   to use or free it.

 - Make all field accesses to WASM-created structs go through accessor
   functions on the WASM side

This also updates the Makefile to make the debug and release targets
write to the same output wasm file, which is convenient for switching
between them when testing.